Move new tests to separate file
